Add ExUnit compiler
Vim compilers (:h compilers) allow vim to run a compiler-like program
(:h 'makeprg'), parse errors in the output (:h 'errorformat'), and display
them in the QuickFix list (:h quickfix). This adds a compiler for ExUnit
that runs `mix test` and parses the output to display test failures. It
can be used as:

    :compiler exunit
    :make

TODO: Formatter currently only supports failures, not exceptions.