Updating the error format to process: * `** (SyntaxError) file:# error message` * `  #) test test unit fail message` * process the multi-line failures from ExUnit and highlight each file
Shortcomings at the moment is there is not a good way to distinguish
local errors vs dependency errors. This can be enhanced by changing the
output formatters to intelligent detect the local app name. For now all
the files are captured and it allows jumping to file from quickfix list,
but obviously erlang libs go to an empty new file