fix indentation when following previous end block
The changes to limit the number of lines we search for a block have made
it so that `def` calls in modules don't indent properly. This is because
we're looking for the base module declaration as a reference point and
stop searching before we get there

Instead, we hack around a bit and try to detect this by looking for
the last `end`. This solution isn't perfect, but it's cheap and effective
until we find problems with it