Use 'setfiletype' to set local &filetype
'setfiletype' is nice in these cases because it prevents &filetype from
being set redundantly.

Note that the Vim runtime ships with a filetype match for *.ex (for
Euphoria), but our plugin-based filetype detection still gets the last
say.

Also, use 'setlocal' when setting &filetype in the shebang detection
path. We don't want to use 'setfiletype' here because it's possible we
need to override an existing syntax assignment, and we already guard
against setting &filetype to 'elixir' twice.