Remove ftdetect
Elixir and all of the most popular template syntaxes
(LEEx, HEEx, Surface) are included in Vim and Neovim's native filetype
detection and now longer have to provided by a plugin.

I noticed the FileType event was being emitted twice when opening Elixir
files, which was causing me some problems. This should fix that.

This also switches the .lexs extension for eelixir to elixir, as that
those files are have elixir in them, not eelixir.