Fix block start pattern
The regex pattern used to discover a block when it spotted a `with`.
This lead to the following code:

```elixir
defmodule Broken do
  def test do
    f(:arg1, with: :arg2)
  end # ...when we hit Enter/Return here...
end
```

Being improperly indented:

```elixir
defmodule Broken do
  def test do
    f(:arg1, with: :arg2)
      end
end
```

The more `with`s, the merrier:

```elixir
defmodule Broken do
  def test do
    f(:arg1, with: :arg2)
    |> g(with: :arg3)
  end

      def test2 do # with 2 `with`s this line got indented like this

  end
end
```

This commit uses a negative lookahead to make sure the block start pattern
only matches if the keyword is not immediately followed
by a `:` colon character.